`profile` in Accept header must be a quoted-string
I discovered recently that the usage of the `profile` Accept header extension
described in the documentation is not a valid accept header extension.

```
Content-Type: application/json; profile=/schema-for-this-data
```

The schema should be within quotes in order the valid.
```
Content-Type: application/json; profile="/schema-for-this-data"
```

This is how the `Content-Type` header is defined
```
Content-Type   = "Content-Type" ":" media-type

media-type     = type "/" subtype *( ";" parameter )
type           = token
subtype        = token

parameter      = attribute "=" value
attribute      = token
value          = token | quoted-string

token          = 1*<any CHAR except CTLs or separators>
separators     = "(" | ")" | "<" | ">" | "@"
               | "," | ";" | ":" | "\" | <">
               | "/" | "[" | "]" | "?" | "="
               | "{" | "}" | SP | HT
```
http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.17
http://www.w3.org/Protocols/rfc2616/rfc2616-sec3.html#sec3.7
http://www.w3.org/Protocols/rfc2616/rfc2616-sec3.html#sec3.6
http://www.w3.org/Protocols/rfc2616/rfc2616-sec2.html#sec2.2

Therefore, `/schema-for-this-data` is not a valid token because it contains a
`/`.  It needs to be expressed as a `quoted-string`.  This occurs anywhere
`profile` appears in the documentation.