Fix parent link example.
This addresses issue #97, where it was noted that "The only
interpretation that makes sense is it's providing an array
of instances, each can be described by the schema."

This makes the schema an array schema, and replaces the "self"
link with the more correct (or at least less confusing in this
context) "item" link.

Also fix a "the the" typo in the security considerations section,
and make "hyper-schema" consistent.