Clarify "targetSchema" and HTTP
"targetSchema" is often misinterpreted as a response schema, when
it in fact is the representation of the target resource which may
or may not appear in various responses.  The previous draft
clarified this somewhat by referencing GET responses and PUT
requests.  This separates the HTTP information from the field's
primary definition, and expands it to the other relevant
HTTP cases.