Set up a blank change log for the next draft.
None of the open PRs have a change log, because they would
all conflict with each other.  This initializes empty change
logs and updates the draft numbering accordingly.

Note that there must be at least an empty <t></t> in the list
or the XML won't validate against its DTD.

Ensuring that the entire changelog is filled out is part of
the pre-publication checklist.