Move media from Hyper-Schema as contentMediaType/contentEncoding in Validation
Definition of "media" property (an object in Hyper-Schema) is moved into
two properties "contentMediaType" and "contentEncoding" in the
validation specification.

This addresses the other part of #363, which states that some keywords
historically in the Hyper-Schema specification would be better in
the Validation document. The main argument for moving keywords such as
"media" is that the Hyper-Schema document could then only focus on
describing the hypermedia linking model.

The new keywords are described in a dedicated section "String-encoding
non-JSON data". Part of the previous content was moved into a forewords
section.