Wording improvements from PR review
Define the condition for using hrefSchema more clearly (I hope).

Use consistent language around "locations" in a schema rather
than introduce a new term "data sources" which is not defined.

Attempt a better explanation of how to find values in the instance.
I'm not entirely sure it's successful but at least it's different.