Move HTTP up between Implementation and Examples
While it is technically not a primary part of the specification,
but rather an application of the spec to a specific protocol,
it is important enough to explaining the motivation and usage
of Hyper-Schema to discuss immediately after the general
implementation requirements but before the examples.

This allows the examples to reference and illustrate HTTP usage.