xref for "self" rel type, improve guidance wording
For some reason the reference to the Atom RFC, where "self" was
originally defined, was commented out.  Restore the reference
and cite it.

The text for "self" was confusing in that it was unclear which
parts applied in general and which parts applied to only "self"
links for the entire instance document.  Rearrange some sentences
to make that more clear.