Name the split "dependencies" to "dependent*"
Upon further discussion, the grammatical approach that
is most intuitive was deemed to be that the string array
or schema's usage is dependent upon the presence or
absence of the instance property.

In at least some anecdotal reports, the nature and
direction of the "dependencies" concept proved confusing.
Others found it intuitive.  The revised wording aims
to explicitly tie the dependency direction to the
naming convention.

In addition, while there was some debate over the most
migration-friendly strategy, the decision is to leave
the old form of "dependencies" available as an extension,
and for at least one draft use the meta-schema to avoid
its redefinition.  Therefore both keywords resulting
from the split have been renamed.

"dependentSchemas" was also in the wrong section.  As an
in-place conditional applicator, it belongs with "if", "then",
and "else".  Not with the object child applicators.

Apparently, I had not made the meta-schema changes at all.
The "$comment" follows the precedent established from leaving
"definitions" in the meta-schema after the rename to "$defs".